+++ tests/PSRule.Rules.Azure.Tests/Arm/Expressions/ExpressionBuilderTests.cs
@@ -149,6 +149,16 @@ public sealed class ExpressionBuilderTests
         Assert.NotNull(actual);
     }
 
+    [Fact]
+    public void BuildExpressionWithRoleDefinitions()
+    {
+        var context = GetContext();
+
+        var actual = Build(context, "[roleDefinitions('Contributor').id]") as JValue;
+
+        Assert.Equal("/subscriptions/ffffffff-ffff-ffff-ffff-ffffffffffff/providers/Microsoft.Authorization/roleDefinitions/00000000-0000-0000-0000-000000000000", actual.Value<string>());
+    }
+
     private static object Build(TemplateContext context, string expression)
     {
         var builder = new ExpressionBuilder();