+++ docs/changelog.md
@@ -37,6 +37,8 @@ What's changed since pre-release v1.48.0-B0088:
     [#3790](https://github.com/Azure/PSRule.Rules.Azure/issues/3790)
   - Added support for Bicep `resourceType` parameter metadata.
     [#1474](https://github.com/Azure/PSRule.Rules.Azure/issues/1474)
+  - Added support for the Bicep `roleDefinitions` function.
+    [#3776](https://github.com/Azure/PSRule.Rules.Azure/issues/3776)
 
 ## v1.48.0-B0088 (pre-release)